Generate msg/scene types and ids at compile time instead of manual incrementing

+++ frontend/src/scenes/scene.h
@@ -29,7 +29,7 @@
 namespace ibh {
     class scene {
     public:
-        scene(unsigned int type) : _id(0), _type(type), _closed(false) {}
+        scene(uint64_t type) : _id(0), _type(type), _closed(false) {}
         virtual ~scene() = default;
 
         virtual void update(iscene_manager *manager, TimeDelta dt) = 0;
@@ -45,7 +45,7 @@ namespace ibh {
         }
 
         unsigned int _id;
-        unsigned int _type;
+        uint64_t _type;
         bool _closed;
     };
